Updating to use [-1, 1] when using Legendre polynomials.
There were places (e.g. the ill-conditioning unit test) where
points on [0, 1] were used with Legendre expansions on [-1, 1].
Also updated the docstring and implementation of find_matrices
to be explicit about what is being computed.

This showed that the evenly-spaced points still give an
ill-conditioned Legendre coefficient matrix (for inverting)
but not nearly as bad as previously indicated.